requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
filelock==3.16.1
fsspec==2024.12.0
idna==3.11
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
pillow==11.1.0
python-dotenv==1.2.1
requests==2.32.5
setuptools==75.8.0
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
urllib3==2.6.2